+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## Unreleased
 ### Modified
 - fetching commented and uniq accepted reviews from Github
+- calculating commented PRs
 
 ## [1.7.8] - 2024-05-03
 ### Added